+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 - PRODUCT - added tracking on code and name fields.
 
 ## Improvements
+- FIXES ASSET : add analytic distribution template.
+- FIXES ASSET CATEGORY : add analytic distribution template.
+
 ## Bug Fixes
 - Ebics User : resolve error getting on export & Modify import config and export template to include BankOrderList and BankStatementList of EbicsPartner
 - BankOrder : Fix NPE on click of confirm for International transfer.